Seperate Message list
